Update test case 2 (Capabilities) to support SPDM 1.3
This updates test case 2 to check for and confirm report
of SPDM 1.3 in GET_CAPABILITIES. All tests are the same
at this time, and the test body for 1.2 has been made common
with checks to toggle support for 1.3.

This does not validate any new capabilities, such as the supported
algorithms block, added in SPDM 1.3.

Signed-off-by: Will Marone <Will.Marone@amd.com>